# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for managing todos with JWT authentication",
        "title": "GoTodo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/feed/discover": {
            "get": {
                "description": "Get discovery feed of trending/popular public anchors from non-followed users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get discovery feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category: trending, popular, recent (default trending)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by tag",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/feed.DiscoverResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/feed/following": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get personalized feed of anchors from followed users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get home feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include own anchors (default true)",
                        "name": "includeOwn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/feed.HomeFeedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/feed/tags/{tagName}": {
            "get": {
                "description": "Get public anchors for a specific tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get tag feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag name",
                        "name": "tagName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/feed.DiscoverResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/reports": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "safety"
                ],
                "summary": "Report content or user",
                "parameters": [
                    {
                        "description": "Report details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/safety.CreateReportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/users/me/blocks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of blocked users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "safety"
                ],
                "summary": "Get blocked users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/auth.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/block": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "safety"
                ],
                "summary": "Block a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to block",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "safety"
                ],
                "summary": "Unblock a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to unblock",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.User": {
            "type": "object",
            "properties": {
                "anchorCount": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "blockedUsers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coverImageUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followerCount": {
                    "type": "integer"
                },
                "followingCount": {
                    "type": "integer"
                },
                "googleId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isVerified": {
                    "type": "boolean"
                },
                "joinedAt": {
                    "type": "string"
                },
                "profilePictureUrl": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "usernameChanged": {
                    "type": "boolean"
                },
                "usernameChangedAt": {
                    "type": "string"
                }
            }
        },
...
        "feed.DiscoverItem": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/feed.DiscoverItemAuthor"
                },
                "cloneCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "coverMediaType": {
                    "type": "string"
                },
                "coverMediaValue": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "engagement": {
                    "$ref": "#/definitions/feed.FeedEngagement"
                },
                "engagementScore": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isPinned": {
                    "type": "boolean"
                },
                "itemCount": {
                    "type": "integer"
                },
                "lastItemAddedAt": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "preview": {
                    "$ref": "#/definitions/feed.FeedPreview"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "feed.DiscoverItemAuthor": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "followerCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "profilePicture": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "feed.DiscoverMeta": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "emptyReason": {
                    "type": "string"
                },
                "feedType": {
                    "type": "string"
                },
                "isAuthenticated": {
                    "type": "boolean"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "feed.DiscoverResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.DiscoverItem"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/feed.DiscoverMeta"
                },
                "pagination": {
                    "$ref": "#/definitions/feed.FeedPagination"
                }
            }
        },
        "feed.FeedEngagement": {
            "type": "object",
            "properties": {
                "hasCloned": {
                    "type": "boolean"
                },
                "hasLiked": {
                    "type": "boolean"
                },
                "likeSummary": {
                    "$ref": "#/definitions/feed.FeedLikeSummary"
                }
            }
        },
        "feed.FeedItem": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/feed.FeedItemAuthor"
                },
                "cloneCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "coverMediaType": {
                    "type": "string"
                },
                "coverMediaValue": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "engagement": {
                    "$ref": "#/definitions/feed.FeedEngagement"
                },
                "id": {
                    "type": "string"
                },
                "isPinned": {
                    "type": "boolean"
                },
                "itemCount": {
                    "type": "integer"
                },
                "lastItemAddedAt": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "preview": {
                    "$ref": "#/definitions/feed.FeedPreview"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "feed.FeedMeta": {
            "type": "object",
            "properties": {
                "emptyReason": {
                    "type": "string"
                },
                "feedType": {
                    "type": "string"
                },
                "includesOwnAnchors": {
                    "type": "boolean"
                },
                "totalFollowing": {
                    "type": "integer"
                }
            }
        },
        "feed.FeedPagination": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "itemCount": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "nextCursor": {
                    "type": "string"
                }
            }
        },
        "feed.FeedPreview": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedPreviewItem"
                    }
                }
            }
        },
...
        "feed.FeedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedItem"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/feed.FeedMeta"
                },
                "pagination": {
                    "$ref": "#/definitions/feed.FeedPagination"
                }
            }
        },
        "feed.FollowingAnchorFeedItem": {
            "type": "object",
            "properties": {
                "coverMediaType": {
                    "type": "string"
                },
                "coverMediaValue": {
                    "type": "string"
                },
                "currentVersion": {
                    "type": "integer"
                },
                "hasUpdate": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastSeenVersion": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "feed.FollowingAnchorsSection": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FollowingAnchorFeedItem"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "feed.HomeFeedResponse": {
            "type": "object",
            "properties": {
                "feed": {
                    "$ref": "#/definitions/feed.FeedResponse"
                },
                "followingAnchors": {
                    "$ref": "#/definitions/feed.FollowingAnchorsSection"
                },
                "suggestedCategories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.SuggestedCategory"
                    }
                }
            }
        },
        "feed.SuggestedCategory": {
            "type": "object",
            "properties": {
                "anchorCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH_INVALID_TOKEN"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "safety.CreateReportRequest": {
            "type": "object",
            "required": [
                "reason",
                "targetId",
                "targetType"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 5
                },
                "targetId": {
                    "type": "string"
                },
                "targetType": {
                    "type": "string",
                    "enum": [
                        "anchor",
                        "item",
                        "user"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}